Convert to 2.6.0: Modularize all kinds Enyo-DCO-1.1-Signed-off-by: Stephen Choi <stephen.choi@lge.com>
