+++ .gitignore
@@ -1,2 +1,3 @@
-/build
-/deploy/moon-flickr
+lib
+dist
+.enyocache
