+++ index.js
@@ -1,0 +1,9 @@
+var
+	ready = require("enyo/ready");
+
+var
+	App = require("./src/app");
+
+ready(function () {
+	new App({name: "app"});
+});
