+++ package.json
@@ -1,0 +1,8 @@
+{
+  "name": "moon-flickr",
+  "assets": [],
+  "devAssets": [],
+  "styles": [],
+  "main": "index.js",
+  "moduleDir": "src"
+}
