--- source/package.js
@@ -1,23 +1,0 @@
-enyo.depends(
-	"../enyo/source/2.6.0-pre.4-wip-package.js",
-	// Cordova (PhoneGap) library
-	// "$lib/enyo-cordova",
-	// webOS-specific kind library
-	// "$lib/enyo-webos",
-	// Layout library
-	"$lib/layout",
-	// Internationalization library
-	// "$lib/enyo-ilib",
-	// Moonstone UI library
-	"$lib/moonstone",
-	// Focus support library for TV applications
-	"$lib/spotlight",
-	// CSS/LESS style files
-	"style",
-	// Model and data definitions
-	"data",
-	// Controller kinds for the application(s)
-	"views",
-	// Include our default entry point
-	"app.js"
-);
