+++ src/data/ActionConstants.js
@@ -1,0 +1,3 @@
+module.exports = {
+	search: 'search'
+};
